requirements.txt
# Existing
streamlit>=1.30
anthropic>=0.40
httpx>=0.27
openpyxl>=3.1
pydantic>=2.0
python-docx>=1.0
python-dotenv>=1.0

# Phase 1: FastAPI backend
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.20

# Phase 2: Azure Blob Storage
azure-storage-blob>=12.0

# Phase 3: Azure AD auth
python-jose[cryptography]>=3.3
